python-keycloak-soo/requirements.txt
fastapi == 0.110.1
pydantic == 2.7.4
python-keycloak == 4.3.0
uvicorn == 0.30.1